requirements.txt
discord.py
python-dotenv
flask
google-generativeai